+++ tests/entryProperties.test.ts
@@ -9,7 +9,7 @@ describe("entry properties", () => {
 key: value
 text down here
 `;
-    expect(first(mw).properties).toEqual([["key", "value"]]);
+    expect(first(mw).properties).toEqual({ key: "value" });
   });
 
   test.each(sp())(
@@ -23,10 +23,10 @@ otherKey: value
 - [] another list item
 `;
       const firstEvent = first(mw);
-      expect(firstEvent.properties).toEqual([
-        ["properties", true],
-        ["otherKey", "value"],
-      ]);
+      expect(firstEvent.properties).toEqual({
+        properties: true,
+        otherKey: "value",
+      });
 
       expect(firstEvent.firstLine.restTrimmed).toBe("a year of events");
       expect(firstEvent.supplemental.length).toBe(2);
@@ -50,12 +50,12 @@ old: yes`;
     const second = nthEvent(events, 1);
     const third = nthEvent(events, 2);
 
-    expect(first.properties).toEqual([
-      ["thisHasProperties", [["indented", "some value with spaces"]]],
-    ]);
+    expect(first.properties).toEqual({
+      thisHasProperties: { indented: "some value with spaces" },
+    });
     expect(first.supplemental.length).toBe(0);
 
-    expect(second.properties).toEqual([]);
+    expect(second.properties).toEqual({});
     expect(second.firstLine.restTrimmed).toBe("future time");
   });
 });
@@ -65,7 +65,7 @@ describe("proper ranges", () => {
     const mw = `1999-09: birthday
 propKey: l`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", "l"]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: "l" });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyKey) {
@@ -81,7 +81,7 @@ propKey: l`;
     const mw = `1999-09: birthday
 propKey:`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", null]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: null });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyKey) {
@@ -97,7 +97,7 @@ propKey:`;
     const mw = `1999-09: birthday
 propKey: l`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", "l"]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: "l" });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyValue) {
@@ -113,7 +113,7 @@ propKey: l`;
     const mw = `1999-09: birthday
 propKey: l  1`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", "l  1"]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: "l  1" });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyValue) {
@@ -129,7 +129,7 @@ propKey: l  1`;
     const mw = `1999-09: birthday
 propKey: l  1`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", "l  1"]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: "l  1" });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyKey) {
@@ -145,9 +145,9 @@ propKey: l  1`;
     const mw = `1999-09: birthday
 x: something:with colon `;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([
-      ["x", "something:with colon"],
-    ]);
+    expect(nthEvent(parsed, 0).properties).toEqual({
+      x: "something:with colon",
+    });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyKey) {
@@ -163,9 +163,9 @@ x: something:with colon `;
     const mw = `1999-09: birthday
 x: something:with colon `;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([
-      ["x", "something:with colon"],
-    ]);
+    expect(nthEvent(parsed, 0).properties).toEqual({
+      x: "something:with colon",
+    });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyValue) {
@@ -181,7 +181,7 @@ x: something:with colon `;
     const mw = `1999-09: birthday
   propKey: l`;
     const parsed = parse(mw);
-    expect(nthEvent(parsed, 0).properties).toEqual([["propKey", "l"]]);
+    expect(nthEvent(parsed, 0).properties).toEqual({ propKey: "l" });
     let visited = false;
     for (const range of parsed.ranges) {
       if (range.type === RangeType.PropertyKey) {
@@ -201,7 +201,7 @@ describe("permitted keys", () => {
 
     const events = parse(mw);
     const first = nthEvent(events, 0);
-    expect(first.properties[0][0]).toBe("key-with-dashes");
+    expect(first.properties["key-with-dashes"]).toBe("value");
   });
 
   test("periods", () => {
@@ -210,7 +210,7 @@ describe("permitted keys", () => {
 
     const events = parse(mw);
     const first = nthEvent(events, 0);
-    expect(first.properties[0][0]).toBe("key-with.period");
+    expect(first.properties["key-with.period"]).toBe("value");
   });
 });
 
@@ -221,7 +221,57 @@ describe("hex values", () => {
       nested: #value`;
     const events = parse(mw);
     const first = nthEvent(events, 0);
-    expect(first.properties[0][1][0][1]).toBe(")value");
+    expect(first.properties.key.nested).toBe(")value");
+  });
+});
+
+describe("prop order", () => {
+  test("propOrder is correct 1", () => {
+    const mw = `1999-09: birthday
+
+group Happy events
+someKey: some value
+otherKey: other value
+
+group Sad events
+property: value
+abc: 123
+
+1995: another event`;
+
+    const events = parse(mw);
+    for (const { eventy } of iter(events.events)) {
+      if (!eventy) {
+        break;
+      }
+      if (isEvent(eventy)) {
+        continue;
+      } else {
+        if (eventy.title === "Happy events") {
+          expect(eventy.propOrder).toEqual(["someKey", "otherKey"]);
+        } else if (eventy.title === "Sad events") {
+          expect(eventy.propOrder).toEqual(["property", "abc"]);
+        }
+      }
+    }
+
+    expect(nthEvent(events, 0).firstLine.restTrimmed).toBe("birthday");
+    expect(nthEvent(events, 1).firstLine.restTrimmed).toBe("another event");
+  });
+
+  test("propOrder is correct 2", () => {
+    const mw = `2028: event
+    key:  
+      nested: #value
+    anotherKey: value
+    kf: 
+      nest:
+        ok: yes
+    oh: yes`;
+    const events = parse(mw);
+    const first = nthEvent(events, 0);
+    expect(first.propOrder).toEqual(["key", "anotherKey", "kf", "oh"]);
+    expect(first.properties.kf.nest.ok).toBe("yes");
   });
 });
 
@@ -248,15 +298,15 @@ abc: 123
         continue;
       } else {
         if (eventy.title === "Happy events") {
-          expect(eventy.properties).toEqual([
-            ["someKey", "some value"],
-            ["otherKey", "other value"],
-          ]);
+          expect(eventy.properties).toEqual({
+            someKey: "some value",
+            otherKey: "other value",
+          });
         } else if (eventy.title === "Sad events") {
-          expect(eventy.properties).toEqual([
-            ["property", "value"],
-            ["abc", 123],
-          ]);
+          expect(eventy.properties).toEqual({
+            property: "value",
+            abc: 123,
+          });
         }
       }
     }